pyproject.toml
[project]
name = "advanced-yaml"
version = "0.4.1"
description = "A suite of tools to advance your usage of yaml.  YASL provides schema definition and verification.  YAQL let's you query your yalm as if it were a database.  YARL provides analysis and reporting.  And YATL allows you to transform yaml from one structure (schema) to another."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "astropy>=7.2.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "colorama==0.4.6",
    "coverage==7.10.6",
    "cucumber-expressions==18.0.1",
    "cucumber-tag-expressions==6.2.0",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "markdown-it-py==4.0.0",
    "mkdocs-material>=9.6.22",
    "packaging==25.0",
    "parse==1.20.2",
    "parse-type==0.6.6",
    "pluggy==1.6.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "requests==2.32.5",
    "ruamel-yaml==0.18.15",
    "ruamel-yaml-clib==0.2.12",
    "setuptools>=80.9.0",
    "six==1.17.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "urllib3==2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-cov==7.0.0",
    "behave==1.3.2",
    "ruff==0.13.0",
    "build==1.3.0",
    "toml==0.10.2",
    "mkdocs==1.6.1",
    "mkdocs-material==9.6.22",
]

[project.scripts]
yasl = "yasl.cli:main"
yaql = "yaql.cli:main"
yarl = "yarl.cli:main"
yatl = "yatl.cli:main"

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "pyright>=1.1.407",
]

[tool.ruff.lint]
ignore = ["E501", "UP045"]
select = ["E", "F", "I", "W", "UP", "B"] # A good starting set of rules

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src --cov-report=html --cov-fail-under=75"
pythonpath = ["src"]
testpaths = [
    "tests/yasl",
    "tests/yaql",
    "tests/yarl",
    "tests/yatl",
]

[tool.behave]
paths = ["features/yasl", "features/yaql", "features/yarl", "features/yatl"]
format = ["pretty"]

[tool.uv]
package = true
